4/26/22 Libra Prototype
Collects 500 libra horoscope tweets, gets ~5000 users, and gets (if they are there) up to 1000 of each of those user's tweets and stores them in separate files. Preprocessing and emotional analysis is then performed, which is averaged per user which is then collapsed into a grand total for the star sign Libra.